refactor(#8/user): UserService 에 CQRS 적용
